--- saleor/static/dashboard-next/categories/components/CategoryDeleteImage/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./CategoryDeleteImage";
-export * from "./CategoryDeleteImage";
