+++ saleor/static/dashboard-next/categories/components/CategoryListPage/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./CategoryListPage";
+export * from "./CategoryListPage";
