--- saleor/static/dashboard-next/categories/components/CategoryTable/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./CategoryTable";
-export * from "./CategoryTable";
